requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
pydantic-settings==2.1.0
openai==1.3.7
bs4==0.0.1
lxml==4.9.3
requests==2.31.0
aiofiles==23.2.1

# Document parsing libraries
pdfminer.six==20221105
python-docx==0.8.11
langdetect==1.0.9
spacy==3.7.2
nltk==3.8.1

# AI/ML integration
transformers==4.38.2
scikit-learn==1.3.2